requirements.txt
schedule==0.6.0
numpy==1.19.2
argparse==1.4.0
imutils==0.5.3
dlib==19.18.0
opencv-python==4.2.0.32
scipy==1.4.1
Pillow==8.4.0
